+++ src/main/kotlin/com/atvgstudios/klib/typealiases/Function.kt
@@ -1,0 +1,9 @@
+package com.atvgstudios.klib.typealiases
+
+/**
+ * Type alias for () -> Unit
+ *
+ * @since 0.1.5
+ * @author Nils Rider
+ */
+typealias Function = () -> Unit