+++ src/test/kotlin/Test.kt
@@ -1,9 +1,3 @@
-import com.atvgstudios.klib.extensions.popBegin
-import com.atvgstudios.klib.extensions.popEnd
-import com.atvgstudios.klib.extensions.prepend
 
 fun main(){
-    val x: MutableList<Int> = mutableListOf()
-    println(x.popBegin())
-    println(x.popEnd())
 }